refactor(bot_tools): Remove unused functions and migrate required functions to response_agent
- Removed unused functions related to file modification and git operations in `bot_tools.py`.
- Moved essential functions `get_tracked_repos` and `is_tool_related` from `bot_tools.py` to `response_agent.py`.
- Updated `response_agent.py` to use the functions internally, removing dependency on `bot_tools.py` for these operations.
- Added a cautionary comment in `bot_tools.py` for safe usage by agents.
